rdctl start should not inherit any handles when starting RD
Otherwise those handles might be kept open for a long time,
as Rancher Desktop detaches from the parent process.
