+++ src/go/rdctl/cmd/start.go
@@ -20,14 +20,11 @@ import (
 	"fmt"
 	"io/fs"
 	"os"
-	"os/exec"
 	"path"
 	"runtime"
-	"strings"
 
 	"github.com/rancher-sandbox/rancher-desktop/src/go/rdctl/pkg/directories"
 	"github.com/rancher-sandbox/rancher-desktop/src/go/rdctl/pkg/options/generated"
-	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
@@ -104,30 +101,6 @@ func doStartCommand(cmd *cobra.Command) error {
 		}
 	}
 	return launchApp(applicationPath, commandLineArgs)
-}
-
-func launchApp(applicationPath string, commandLineArgs []string) error {
-	var commandName string
-	var args []string
-
-	if runtime.GOOS == "darwin" {
-		commandName = "/usr/bin/open"
-		args = []string{"-a", applicationPath}
-		if len(commandLineArgs) > 0 {
-			args = append(args, "--args")
-			args = append(args, commandLineArgs...)
-		}
-	} else {
-		commandName = applicationPath
-		args = commandLineArgs
-	}
-	// Include this output because there's a delay before the UI comes up.
-	// Without this line, it might look like the command doesn't work.
-	logrus.Infof("About to launch %s %s ...\n", commandName, strings.Join(args, " "))
-	cmd := exec.Command(commandName, args...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	return cmd.Start()
 }
 
 func moveToParent(fullPath string, numberTimes int) string {
