+++ src/go/rdctl/go.mod
@@ -1,6 +1,6 @@
 module github.com/rancher-sandbox/rancher-desktop/src/go/rdctl
 
-go 1.17
+go 1.18
 
 require (
 	github.com/docker/docker v20.10.22+incompatible
